pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptops"
version = "0.1.0"
description = "Git-style version control for LLM prompts"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "PromptOps Contributors"}
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "click>=8.1.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
promptops = "promptops.cli.main:main"

[tool.setuptools]
packages = ["promptops", "promptops.cli", "promptops.core", "promptops.api", "promptops.eval", "promptops.deploy", "promptops.integrations"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["promptops"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
fail_under = 80